Initial Gradle-based Android project
This will serve as base for an Gradle-based AndroidAnnotations project.

This makes use of gradle-android-plugin to be processed properly as an
Android project.  The current 1.0.0 release does not work with Android
SDK R14 and higher, due to changes in it's build system.  We thus use
1.1.0-SNAPSHOT deployed in a custom Maven repository for it to work.

The androidSignAndAlign configuration is commented out, as we do not
provide a keystore with this commit.

The IDEA plugin is also added to this project, which will be enhanced in
the future to automatically configure AndroidAnnotations in IDEA.  We also
configure the IDEA plugin to download source and JavaDocs automatically,
as well as set the language level to 1.6.

This project isn't really limited to just Gradle+IDEA, but more on that I
haven't used this with Eclipse yet.  Will add it together with configuration
to make it work with AndroidAnnotations.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>