Included AndroidAnnotations 2.2-RC2 as dependency
We define the API AndroidAnnotations library as a compile-time
dependency in Gradle, as we include this in the APK that we
generate.  The annotation processor JAR is added manually as
we will simply reference this in both our IDEA and compile
configurations to perform annotation processing.

Ideally AndroidAnnotations is defined in Gradle only, and will
download both the API and the annotation processor JARs.  The
latter would be set to provided as we do not want to include that
in the APK, and our IDEA and compile configuration would instead
refer to the annotation processor in the classpath.  I haven't
exactly made this work yet :(.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>