Configuration to process annotations on `gradle assemble`
We override compileJava's doFirst method in order to add AndroidAnnotations
to the compile process.  This allows Gradle to generate the annotations before
compiling the rest of the code.

Unfortunately, the generated code is outputted to build/gen instead of simply
the gen folder.  This means that while the annotated classes are processed
correctly during a `gradle assemble`, they do not show up inside IDEA and must
be rebuilt manually to generate.  It should be possible to point the output
folders of both Gradle and IDEA to the same location, but I haven't really
looked into that yet.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>