Fix IntentBuilder startForResult
In a fragment, startActivityForResult must be called on the Fragment and
not on getActivity(), or the fragment's onActivityResult will never be
hit.

Now IntentBuilder may have up to 3 constructors, depending on the
classpath content. According to the constructor chosen,
startActivityForResult will be called on Activity, android.app.Fragment
or android.support.v4.app.Fragment.

From the AA user perspective, this fix is quite transparent, he just
have to call MyActivity_.intent(this), whatever he's in a Fragment or an
Activity, it will work.