Update Guava to 15.0
We want to integrate Parceler, but we ran into a problem. The Guava
version used by Parceler is 15.0, but our pom is declaring the 13.0.1
version. Because of this when using Parceler in conjunction with AA,
the older Guava dependency is resolved, witch results in problems,
since Parceler needs the new classes introduced since 13.0.1.

This commit addresses this issue by simply updating the Guava version to
15.0. This is enough because our code do not uses any classes which are
changed since 13.0.1.