Fix ClassLoader issues with ParameterizedRobolectricTestRunner
The ParameterizedRobolectricTestRunner has a problem: when creating the
test parameter objects, it uses the default ClassLoader, but for the
tests it uses the Robolectric ClassLoader. This results in errors since
objects cannot be used altogether which are originating from different
ClassLoaders.
This commit fixes this issue by creating a workaround class which loads
Robolectric with a fake test, aquire the Robolectric ClassLoader and
creates the test parameter objects with that.