Change Travis link in README.md
The old link pointed to https://travis-ci.org/excilys/androidannotations, which shows the last build. Since the last build can be a PR build, and not always the latest build on the develop branch, this can show misleading information about the project's status. Changed to https://travis-ci.org/excilys/androidannotations/builds, which shows the build history on the develop branch.