Re-add Otto annotation handlers
We copy all non-AA annotations when we override a method. This can help
Otto, because it does not read annotations from parent classes. But if a
method is only annotated with Otto annotation, so we do not override it,
the annotation will not be read Otto and the methods will not be called.
To help that case, we re-add Otto handlers, which override (and
annotate) Otto annotated methods in the generated subclass.