Fixed classpath issues with RoboGuice
Some classes of RoboGuice are referencing the android.support package.
This is problematic, since it is not always in the classpath. In that
case, generated code cannot be compiled if support classes are
indirectly referenced, also validation can go wrong.
This commit fixes that issue by removing all indirect reference to the
support package.