Commit functional test JDT Eclipse prefs
Unfortunetaly this file is needed because annotation processing is
turned off in Eclipse by default. However, m2e will update this file
when the project is imported. This makes working with git inconvenient,
since we always have to stash the file when changing branching etc. So
we just commit the whole file now, since it does not really make a
difference since the file is already present in the repository.