Do not extract annotations on Android classes
Currently there is a serious bug [1] in the Android Support library,
which is causing processing to fail when extracting annotations from
ancestor classes. The library was compiled with bad options, so some
annotations are missing from the jar. This commit is a workaround for
the problem: we simply skip the extraction when we encounter a class
starting with the "android" package name prefix. This workaround can be
permanent, because this will not effect usage, and also adds a little
optimizations, as AA annotations will not be found ever on Android
classes.

1: https://code.google.com/p/android/issues/detail?id=175086