Support Instant Run application class
Instant Run swaps the application class name in the manifest to its own
bootstrap application, which can instrument the app to use instant Run.
It also adds a new "name" attribute to the manifest, which contains the
original application. When Instant Run is used, we should retrieve this
attribute as the application class name.