Support synchronous executors
BackgroundExecutor assumed that its executor was asynchronous, so that
in execute(Task), a task could not be removed before it has been added
to the TASKS list.

However, this assumption was wrong with a synchronous executor:
directExecute(…) would execute the task synchronously, leading to call
Task#postExecute(), that removes the task from TASKS. Immediately after
the execution, directExecute(…) would add the task to TASKS, which let
the BackgroundExecutor in an incorrect state.

Therefore, add the task to the TASKS list before the execution, so that
it will also work with synchronous executors.

Fixes #1689.
<https://github.com/excilys/androidannotations/issues/1689>