Support R2 class for resource values
The R class contains non-constant fields in case of library projects,
so we cannot use them in annotation values, nor for validaiton, since
the values are not available at compile time. To work around this,
there is a Gradle plugin to generate an R2 class, which contains
constants fields.

This commit adds a very simple integration to this plugin, by using the
R2 class instead of R if the former is available.